+++ packages/client/ui/react-ui/src/providers/CrossmintAuthProvider.tsx
@@ -1,27 +1,31 @@
 import { useCrossmint } from "@/hooks";
-import { ReactNode } from "react";
+import type { ReactNode } from "react";
 
 import { AuthProvider as AuthCoreProvider } from "@crossmint/client-sdk-auth-core/client";
+import { EVMSmartWalletChain } from "@crossmint/client-sdk-smart-wallet";
+import type { UIConfig } from "@crossmint/common-sdk-base";
 
 import { CrossmintWalletProvider } from "./CrossmintWalletProvider";
 
 export type CrossmintAuthWalletConfig = {
     type: "evm-smart-wallet";
-    defaultChain: "polygon-amoy" | "base-sepolia" | "optimism-sepolia" | "arbitrum-sepolia";
+    defaultChain: EVMSmartWalletChain;
     createOnLogin: "all-users" | "off";
 };
 
 export function CrossmintAuthProvider({
     embeddedWallets,
     children,
+    appearance,
 }: {
     embeddedWallets: CrossmintAuthWalletConfig;
     children: ReactNode;
+    appearance: UIConfig;
 }) {
     const { crossmint, setJwt } = useCrossmint("CrossmintAuthProvider must be used within CrossmintProvider");
 
     return (
-        <AuthCoreProvider setJwtToken={setJwt} crossmint={crossmint}>
+        <AuthCoreProvider setJwtToken={setJwt} crossmint={crossmint} appearance={appearance}>
             <CrossmintWalletProvider
                 walletConfig={{ type: "evm-smart-wallet", signer: { type: "PASSKEY" } }}
                 defaultChain={embeddedWallets.defaultChain}