+++ packages/client/wallets/smart-wallet/src/index.ts
@@ -11,6 +11,8 @@ export type {
     WalletParams,
 } from "./types/params";
 
+export type { SmartWalletChain as EVMSmartWalletChain } from "./blockchain/chains";
+
 export type { TransferType, ERC20TransferType, NFTTransferType, SFTTransferType } from "./types/token";
 
 export {