+++ apps/payments/nextjs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @crossmint/client-sdk-nextjs-starter
 
+## 1.2.16
+
+### Patch Changes
+
+-   Updated dependencies [3865983]
+    -   @crossmint/client-sdk-react-ui@1.3.13
+    -   @crossmint/client-sdk-base@1.2.5
+
 ## 1.2.15
 
 ### Patch Changes