+++ packages/client/ui/vue-ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @crossmint/client-sdk-vue-ui
 
+## 1.1.15
+
+### Patch Changes
+
+-   @crossmint/client-sdk-base@1.2.5
+
 ## 1.1.14
 
 ### Patch Changes